pick generate type definition code
